ref: lumos ai feature has been unified and moved in a dedicated sidebar
