+++ src/rendered/components/navbar/NavDrawer.vue
@@ -136,7 +136,7 @@
     
     /* Sidebar container */
     .sidebar-box {
-        height: calc(100vh - 74px);
+        height: calc(100vh - 60px);
         display: flex;
         flex-direction: column;
         overflow: hidden; /* keep rounded corners when scrolling */
