+++ src/rendered/components/navbar/PageRouter.vue
@@ -16,23 +16,6 @@
     </template>
 </v-list-item>
 
-<!-- AI menu -->
-<v-list-item
-:active="currentRoute.name === 'lumos-ai'"
-class="ma-1"
-rounded="lg"
-@click="openLumosAI()"
->
-<template v-slot:prepend>
-    <v-icon 
-    :icon="currentRoute.name === 'lumos-ai' ? 'mdi-lightning-bolt' : 'mdi-lightning-bolt-outline'"
-    ></v-icon>
-</template>
-<template v-slot:title>
-    <span :class="{ 'font-weight-bold': currentRoute.name === 'lumos-ai' }">Lumos AI</span>
-</template>
-</v-list-item>
-
 <!-- Settings menu -->
 <v-list-item
 :active="currentRoute.name === 'settings'"
@@ -63,9 +46,5 @@ rounded="lg"
     
     const openSettings = () => {
         router.push({ name: 'settings' })
-    }
-    
-    const openLumosAI = () => {
-        router.push({ name: 'lumos-ai' })
     }
 </script>
