+++ src/rendered/stores/aiPreferencesStore.js
@@ -33,14 +33,16 @@ export const aiPreferencesStore = defineStore('aiPreferences', {
         availableModels: {
             ollama: [],
             groq: [
+                'moonshotai/kimi-k2-instruct',
+                'moonshotai/kimi-k2-instruct-0905',
+                'openai/gpt-oss-120b',
+                'openai/gpt-oss-20b',
                 'llama-3.1-8b-instant',
                 'llama3-70b-8192',
-                'llama3-8b-8192',
-                'llama-3.3-70b-versatile',
                 'meta-llama/llama-4-scout-17b-16e-instruct',
                 'meta-llama/llama-4-maverick-17b-128e-instruct',
-                'deepseek-r1-distill-llama-70b',
-                'gemma2-9b-it',
+                'groq/compound',
+                'groq/compound-mini',
                 'qwen-qwq-32b'
             ],
             openai: [
